work around to rename directory to lower case
