+++ src/app-components/LocationDetail2/location-detail.css
